requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
crewai==0.80.0
crewai-tools==0.14.0
pandas==2.2.3
openpyxl==3.1.5
plotly==5.24.1
python-multipart==0.0.12
python-dotenv==1.0.1
websockets==13.1
langchain-community==0.3.0
